requirements.txt
flask
flask-pymongo
bcrypt
pyotp
pyjwt
dnspython
